Convert user favorite songs to React
